spring升级

+++ exam/src/main/java/com/virgo/exam/vo/PublishExamPaperVO.java
@@ -44,7 +44,7 @@ public class PublishExamPaperVO {
     /**
      * 平均难度
      */
-    private Integer avgDifficult;   //平均难度
+    private Float avgDifficult;   //平均难度
     /**
      * 考试时长
      */
